+++ server/database/sqls/Insertsalesorder.js
@@ -1,6 +1,6 @@
 // server/database/sqls/Insertsalesorder.js
 module.exports = {
-
+  // 주문서 등록
   insertSalesOrder: 
   `INSERT INTO salesorder (
     sorder_code, 
@@ -12,5 +12,22 @@ module.exports = {
     emp_id
   ) VALUES (?, ?, ?, ?, ?, ?, ?)`,
 
+  // 거래처 목록 조회
+  selectClientList: 
+  `SELECT c.client_code,
+    c.client_name,
+    c.client_mgr,
+    com.code_name
+  FROM client as c
+  JOIN common_codes as com
+    ON c.code_values = com.code_values
+  WHERE com.code_values = 'E02'
+  ORDER BY c.client_code`,
 
+  // 품목 목록 조회
+  selectItemList: 
+  `SELECT item_code,
+    item_name,
+    item_type
+  FROM items`,
 }
