+++ server/database/sqls/salesorder.js
@@ -165,24 +165,8 @@ module.exports = {
 
   // 주문서 삭제
   deleteSalesOrder: 
-  `DELETE FROM salesorder
-    WHERE sorder_code = ?`,
+  `DELETE 
+  FROM salesorder
+  WHERE sorder_code = ?`,
 
-  selectClientList: 
-  `SELECT c.client_code,
-    c.client_name,
-    c.client_mgr,
-    com.code_name
-  FROM client as c
-  JOIN common_codes as com
-    ON c.code_values = com.code_values
-  WHERE com.code_values = 'E02'
-  ORDER BY c.client_code`,
-
-  selectItemList: 
-  `SELECT item_code,
-    item_name,
-    item_type
-  FROM items`,
-    
 };
