+++ server/routers/Insertsalesorder_router.js
@@ -16,4 +16,24 @@ router.post('/salesorders', async (req, res) => {
     }
 });
 
+// 거래처 목록 조회
+router.get('/clients', async (req, res) => {
+    try {
+        const clients = await isoService.findAllClients();
+        res.send(clients);
+    } catch (err) {
+        res.status(500).send({ error: '거래처 목록을 불러오는데 실패했습니다.' });
+    }
+});
+
+// 품목 목록 조회
+router.get('/items', async (req, res) => {
+    try {
+        const items = await isoService.findAllItems();
+        res.send(items);
+    } catch (err) {
+        res.status(500).send({ error: '품목 목록을 불러오는데 실패했습니다.' });
+    }
+});
+
 module.exports = router;
